+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## [1.2.9] - September 30th, 2020
+## [1.2.9] - October 3rd, 2020
 
 * Added ability to register a custom route to the `TestStepRegistry` and have that route be presented as an option on `TestStepsDialog` and `TestStepsPage`.
 * Added variable for `_now` to return the current `DateTime`.
+* Fixed bug in `TestStepForm` in the dropdown form when values aren't strings.
+* Beginning of external test driver capabilities
 
 
 ## [1.2.8+2] - September 29th, 2020